refactor: move CSV regression analysis from caliper engine to rhaiis plugin
Move compare_runs, run_regression_analysis, RegressionResult, and
geometric_mean out of the generic caliper/engine/kpi/analyze.py into
rhaiis/postprocess/regression.py. The caliper engine should contain
no project-specific code (no references to model, accelerator, TP,
tokens, etc.).

This is a temporary transition — these CSV-based comparison functions
will be replaced by generic Caliper KPI-based analysis once the
framework lands.
